+++ mods/climate_modpack/weather_pack/weathers/heavy_rain.lua
@@ -318,7 +318,7 @@ local add_close_range_rain_particle = function(player)
 	if hw_utils.is_outdoor(random_pos, rain_texture_size_offset_y) then
 		local wx, wz = 0, 0
 		if breasy then
-			local w = breasy.get_wind(random_pos)
+			local w = hw_utils.get_wind(random_pos)
 			wx = w.x
 			wz = w.z
 		end
@@ -345,7 +345,7 @@ local add_wide_range_rain_particle = function(player)
 	if hw_utils.is_outdoor(random_pos) then
 		local wx, wz = 0, 0
 		if breasy then
-			local w = breasy.get_wind(random_pos)
+			local w = hw_utils.get_wind(random_pos)
 			wx = w.x
 			wz = w.z
 		end