+++ mods/climate_modpack/weather_pack/weathers/rain.lua
@@ -134,7 +134,7 @@ local add_rain_particle = function(player)
 	if hw_utils.is_outdoor(random_pos) then
 		local wx, wz = 0, 0
 		if breasy then
-			local w = breasy.get_wind(random_pos)
+			local w = hw_utils.get_wind(random_pos)
 			wx = w.x
 			wz = w.z
 		end