+++ mods/climate_modpack/weather_pack/weathers/severe_thunderstorm.lua
@@ -567,7 +567,7 @@ severe_thunderstorm.render = function(dtime, player)
 				if hw_utils.is_outdoor(rpos) then
 					local wx, wz = 0, 0
 					if breasy then
-						local w = breasy.get_wind(rpos)
+						local w = hw_utils.get_wind(rpos)
 						wx = w.x * 0.5
 						wz = w.z * 0.5
 					end
@@ -603,7 +603,7 @@ severe_thunderstorm.render = function(dtime, player)
 				if hw_utils.is_outdoor(rpos) then
 					local wx, wz = 0, 0
 					if breasy then
-						local w = breasy.get_wind(rpos)
+						local w = hw_utils.get_wind(rpos)
 						wx = w.x
 						wz = w.z
 					end
@@ -642,7 +642,7 @@ severe_thunderstorm.render = function(dtime, player)
 				if hw_utils.is_outdoor(rpos) then
 					local wx, wz = 0, 0
 					if breasy then
-						local w = breasy.get_wind(rpos)
+						local w = hw_utils.get_wind(rpos)
 						wx = w.x * 0.5
 						wz = w.z * 0.5
 					end